+++ lib/index.ts
@@ -7,4 +7,4 @@ export { Router, Handler } from "./middleware/router";
 export { apply } from "./middleware/apply";
 export { compose } from "./middleware/compose";
 
-export {Middleware, MiddlewareNextFunction} from "./middlewareTypes";
+export { Middleware, MiddlewareNextFunction } from "./middlewareTypes";