+++ lib/middleware/apply.ts
@@ -5,9 +5,9 @@ import { compose } from "./compose";
 /**
  *  Apply the given middleware to the context, useful for testing
  */
-export function apply(
-  middleware: Middleware | Middleware[],
-  ctx: Context,
+export function apply<EnvironmentType>(
+  middleware: Middleware<EnvironmentType> | Middleware<EnvironmentType>[],
+  ctx: Context<EnvironmentType>,
 ) {
   if (Array.isArray(middleware)) {
     middleware = compose(middleware);